Rename AI Provider to AI Client
