+++ gitbutler-ui/src/lib/backend/aiClient.ts
@@ -1,0 +1,13 @@
+export enum MessageRole {
+	User = 'user',
+	Assistant = 'assisstant'
+}
+
+export interface PromptMessage {
+	content: string;
+	role: MessageRole;
+}
+
+export interface AIClient {
+	evaluate(prompt: string): Promise<string>;
+}
