+++ gitbutler-ui/src/lib/backend/aiClients/anthropic.ts
@@ -1,0 +1,34 @@
+import { type AIClient, type PromptMessage, MessageRole } from '$lib/backend/aiClient';
+import { fetch, Body } from '@tauri-apps/api/http';
+import type { AnthropicModel } from '$lib/backend/aiService';
+
+type AnthropicAPIResponse = { content: { text: string }[] };
+
+export class AnthropicAIClient implements AIClient {
+	constructor(
+		private apiKey: string,
+		private model: AnthropicModel
+	) {}
+
+	async evaluate(prompt: string) {
+		const messages: PromptMessage[] = [{ role: MessageRole.User, content: prompt }];
+
+		const body = Body.json({
+			messages,
+			max_tokens: 1024,
+			model: this.model
+		});
+
+		const response = await fetch<AnthropicAPIResponse>('https://api.anthropic.com/v1/messages', {
+			method: 'POST',
+			headers: {
+				'x-api-key': this.apiKey,
+				'anthropic-version': '2023-06-01',
+				'content-type': 'application/json'
+			},
+			body
+		});
+
+		return response.data.content[0].text;
+	}
+}
